+++ copper-server-impl/src/main/java/com/pogeyan/cmis/impl/factory/DatabaseServiceFactory.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package com.pogeyan.cmis.service.factory;
+package com.pogeyan.cmis.impl.factory;
 
 import java.util.HashMap;
 import java.util.Map;
