+++ copper-server-impl/src/main/java/com/pogeyan/cmis/impl/storage/FileSystemStorageService.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package com.pogeyan.cmis.document.storage;
+package com.pogeyan.cmis.impl.storage;
 
 import java.io.File;
 import java.io.FileInputStream;
