+++ copper-server/src/main/java/com/pogeyan/cmis/actors/AclActor.java
@@ -32,7 +32,7 @@ import com.pogeyan.cmis.api.messages.CmisBaseResponse;
 import com.pogeyan.cmis.api.messages.PostRequest;
 import com.pogeyan.cmis.api.messages.QueryGetRequest;
 import com.pogeyan.cmis.api.utils.Helpers;
-import com.pogeyan.cmis.services.CmisAclServices;
+import com.pogeyan.cmis.impl.services.CmisAclServices;
 
 public class AclActor extends BaseClusterActor<BaseRequest, BaseResponse> {
