+++ copper-server/src/main/java/com/pogeyan/cmis/actors/ObjectActor.java
@@ -63,13 +63,13 @@ import com.pogeyan.cmis.api.messages.PostRequest;
 import com.pogeyan.cmis.api.messages.QueryGetRequest;
 import com.pogeyan.cmis.api.utils.Helpers;
 import com.pogeyan.cmis.api.utils.MetricsInputs;
+import com.pogeyan.cmis.impl.services.CmisObjectService;
+import com.pogeyan.cmis.impl.services.CmisTypeCacheService;
+import com.pogeyan.cmis.impl.services.CmisVersioningServices;
+import com.pogeyan.cmis.impl.utils.CmisPropertyConverter;
+import com.pogeyan.cmis.impl.utils.CmisUtils;
+import com.pogeyan.cmis.impl.utils.DBUtils;
 import com.pogeyan.cmis.api.data.IBaseObject;
-import com.pogeyan.cmis.services.CmisObjectService;
-import com.pogeyan.cmis.services.CmisTypeCacheService;
-import com.pogeyan.cmis.services.CmisVersioningServices;
-import com.pogeyan.cmis.utils.CmisPropertyConverter;
-import com.pogeyan.cmis.utils.CmisUtils;
-import com.pogeyan.cmis.utils.DBUtils;
 
 public class ObjectActor extends BaseClusterActor<BaseRequest, BaseResponse> {
 	private static final Logger LOG = LoggerFactory.getLogger(ObjectActor.class);
