+++ copper-server/src/main/java/com/pogeyan/cmis/actors/PolicyActor.java
@@ -35,8 +35,8 @@ import com.pogeyan.cmis.api.messages.CmisBaseResponse;
 import com.pogeyan.cmis.api.messages.PostRequest;
 import com.pogeyan.cmis.api.messages.QueryGetRequest;
 import com.pogeyan.cmis.api.utils.Helpers;
-import com.pogeyan.cmis.services.CmisObjectService;
-import com.pogeyan.cmis.services.CmisPolicyService;
+import com.pogeyan.cmis.impl.services.CmisObjectService;
+import com.pogeyan.cmis.impl.services.CmisPolicyService;
 
 public class PolicyActor extends BaseClusterActor<BaseRequest, BaseResponse> {
