+++ copper-server/src/main/java/com/pogeyan/cmis/actors/RelationshipActor.java
@@ -33,7 +33,7 @@ import com.pogeyan.cmis.api.BaseResponse;
 import com.pogeyan.cmis.api.messages.CmisBaseResponse;
 import com.pogeyan.cmis.api.messages.QueryGetRequest;
 import com.pogeyan.cmis.api.utils.Helpers;
-import com.pogeyan.cmis.services.CmisRelationshipService;
+import com.pogeyan.cmis.impl.services.CmisRelationshipService;
 
 public class RelationshipActor extends BaseClusterActor<BaseRequest, BaseResponse> {
